iOS: move device.rb to ios/device dir
