+++ lib/calabash/ios.rb
@@ -26,7 +26,7 @@ module Calabash
 
     require 'calabash/ios/environment'
     require 'calabash/ios/physical_device_mixin'
-    require 'calabash/ios/device'
+    require 'calabash/ios/device/device'
     require 'calabash/ios/operations'
     require 'calabash/ios/server'
     require 'calabash/ios/application'
