+++ lib/calabash/ios/device/device.rb
