+++ spec/integration/ios/device/device_spec.rb
