+++ spec/lib/ios/device/device_spec.rb
